pyproject.toml
[project]
name = "hantang-api"
version = "0.0.1"
description = "Default template for PDM package"
authors = [
    {name = "Ovler", email = "ovlertheyoung@gmail.com"},
]
dependencies = ["SQLModel>=0.0.22", "fastapi[standard]>=0.115.6", "pre-commit>=4.0.1", "python-dotenv>=1.0.1", "sqlalchemy>=2.0.36", "asyncpg>=0.30.0", "requests>=2.32.3"]
requires-python = "==3.12.*"
readme = "README.md"
license = {text = "MIT"}


[tool.pdm]
distribution = false
